add test suites
