+++ .github/workflows/tests.yaml
@@ -39,7 +39,7 @@ jobs:
 
       - name: Install pip packages
         run: |
-          pip install --no-cache-dir .[strict,dev,fireworks,optimizers,parsl] # remove --no-cache-dir after repos are removed from setup.py
+          pip install .[strict,dev] # remove  after repos are removed from setup.py
           pip install covalent-slurm-plugin
 
       - name: Configure Quacc
@@ -87,7 +87,7 @@ jobs:
 
       - name: Install pip packages
         run: |
-          pip install --no-cache-dir .[strict,dev,tblite] # remove --no-cache-dir after repos are removed from setup.py
+          pip install .[strict,dev,tblite] # remove  after repos are removed from setup.py
 
       - name: Configure Quacc
         run: quacc config
@@ -132,7 +132,7 @@ jobs:
 
       - name: Install pip packages
         run: |
-          pip install --no-cache-dir .[strict,dev] # remove --no-cache-dir after repos are removed from setup.py
+          pip install .[strict,dev] # remove  after repos are removed from setup.py
 
       - name: Configure Quacc
         run: quacc config
@@ -181,7 +181,7 @@ jobs:
 
       - name: Install pip packages
         run: |
-          pip install --no-cache-dir .[strict,dev] # remove --no-cache-dir after repos are removed from setup.py
+          pip install .[strict,dev] # remove  after repos are removed from setup.py
 
       - name: Configure Quacc
         run: quacc config
@@ -200,10 +200,58 @@ jobs:
           token: ${{ secrets.CODECOV_TOKEN }}
           files: ./coverage.xml
 
+  tests-parsl:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: true
+      matrix:
+        python-version: ["3.10"]
+
+    defaults:
+      run:
+        shell: bash -l {0}
+
+    steps:
+      - name: Check out repo
+        uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: pip
+          cache-dependency-path: setup.py
+
+      - name: Install pip packages
+        run: |
+          pip install .[strict,dev,fireworks,optimizers,parsl]
+          pip install covalent-slurm-plugin
+
+      - name: Configure Quacc
+        run: quacc config
+
+      - name: Set up Covalent server
+        run: |
+          echo "y" | covalent purge
+          covalent start
+
+      - name: Run tests with pytest
+        run: pytest tests/fireworks_tests tests/parsl_tests tests/optimizers --cov=quacc --cov-report=xml
+
+      - name: Stop Covalent server
+        run: covalent stop
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v3
+        if: github.repository == 'quantum-accelerators/quacc'
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: ./coverage.xml
+
   release:
     runs-on: ubuntu-latest
     needs: [tests-base, tests-tblite, tests-psi4, tests-dftbplus]
-    if: github.event_name == 'release' && needs.tests-base.result == 'success' && needs.tests-tblite.result == 'success' && needs.tests-psi4.result == 'success' && needs.tests-dftbplus.result == 'success'
+    if: github.event_name == 'release' && needs.tests-base.result == 'success'
     steps:
       - name: Check out repo
         uses: actions/checkout@v3
