+++ tests/util/test_calc.py
@@ -11,11 +11,6 @@ from ase.optimize import BFGS, BFGSLineSearch
 
 from quacc.util.calc import run_ase_opt, run_ase_vib, run_calc
 
-try:
-    import sella
-except ImportError:
-    sella = None
-
 CWD = os.getcwd()
 
 
@@ -157,43 +152,6 @@ def test_run_ase_opt():
             scratch_dir="test_calc",
             copy_files=["test_file.txt"],
         )
-
-
-@pytest.mark.skipif(
-    sella is None,
-    reason="Sella must be installed.",
-)
-def test_sella():
-    from sella.optimize import Sella
-
-    atoms = bulk("Cu") * (2, 1, 1)
-    atoms[0].position += 0.1
-    atoms.calc = EMT()
-    dyn = run_ase_opt(
-        atoms,
-        optimizer=Sella,
-        scratch_dir="test_calc",
-        gzip=False,
-        copy_files=["test_file.txt"],
-        optimizer_kwargs={"restart": None},
-    )
-    traj = dyn.traj
-    assert traj[-1].calc.results is not None
-    assert dyn.user_internal is False
-
-    atoms = molecule("H2O")
-    atoms.calc = LennardJones()
-    dyn = run_ase_opt(
-        atoms,
-        optimizer=Sella,
-        scratch_dir="test_calc2",
-        gzip=False,
-        copy_files=["test_file.txt"],
-        optimizer_kwargs={"restart": None},
-    )
-    traj = dyn.traj
-    assert traj[-1].calc.results is not None
-    assert dyn.user_internal is True
 
 
 def test_run_ase_vib():
